Flatten files again
