--- app/src/main/java/com/imashnake/animite/features/media/MediaGenres.kt
@@ -1,57 +1,0 @@
-package com.imashnake.animite.features.media
-
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyRow
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.SuggestionChip
-import androidx.compose.material3.SuggestionChipDefaults
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.dimensionResource
-import com.imashnake.animite.R
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun MediaGenres(
-    genres: List<String>,
-    modifier: Modifier = Modifier,
-    contentPadding: PaddingValues,
-    color: Color = MaterialTheme.colorScheme.primaryContainer
-) {
-    LazyRow(
-        horizontalArrangement = Arrangement.spacedBy(
-            dimensionResource(R.dimen.medium_padding)
-        ),
-        contentPadding = contentPadding
-    ) {
-        items(genres) { genre ->
-            SuggestionChip(
-                label = {
-                    Text(
-                        text = genre.lowercase(),
-                        color = MaterialTheme.colorScheme.onBackground,
-                        style = MaterialTheme.typography.labelMedium,
-                        modifier = Modifier.padding(
-                            vertical = dimensionResource(R.dimen.small_padding)
-                        )
-                    )
-                },
-                onClick = { },
-                shape = CircleShape,
-                colors = SuggestionChipDefaults.suggestionChipColors(
-                    containerColor = color.copy(alpha = 0.25f)
-                ),
-                border = SuggestionChipDefaults.suggestionChipBorder(
-                    borderColor = Color.Transparent
-                )
-            )
-        }
-    }
-}
