+++ app/src/main/java/com/imashnake/animite/features/media/MediaPage.kt
@@ -1,14 +1,22 @@
 package com.imashnake.animite.features.media
 
+import android.content.Intent
 import android.content.res.Configuration
+import android.net.Uri
 import android.text.Html
+import android.util.Log
+import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.WindowInsets
 import androidx.compose.foundation.layout.asPaddingValues
+import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.displayCutoutPadding
 import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
@@ -16,23 +24,48 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.navigationBarsPadding
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.statusBars
 import androidx.compose.foundation.layout.statusBarsPadding
 import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.layout.wrapContentSize
+import androidx.compose.foundation.lazy.LazyRow
+import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.verticalScroll
+import androidx.compose.material.ContentAlpha
+import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.SuggestionChip
+import androidx.compose.material3.SuggestionChipDefaults
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.BlendMode
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.ColorFilter
 import androidx.compose.ui.graphics.graphicsLayer
+import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalConfiguration
+import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.res.dimensionResource
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.style.TextOverflow
 import androidx.hilt.navigation.compose.hiltViewModel
+import coil.compose.AsyncImage
+import coil.request.ImageRequest
+import com.imashnake.animite.R
 import com.imashnake.animite.core.extensions.given
+import com.imashnake.animite.core.ui.scrim
+import com.imashnake.animite.dev.internal.Constants
 import com.imashnake.animite.features.ui.MediaSmall
+import com.imashnake.animite.features.ui.MediaSmallRow
 import com.ramcosta.composedestinations.annotation.Destination
 import com.imashnake.animite.R as Res
 
@@ -170,5 +203,228 @@ fun MediaPage(
                 )
                 .align(Alignment.TopCenter)
         ) { }
+    }
+}
+
+@Composable
+fun MediaBanner(
+    imageUrl: String?,
+    tintColor: Color,
+    modifier: Modifier = Modifier
+) {
+    if (!imageUrl.isNullOrEmpty()) {
+        AsyncImage(
+            model = ImageRequest.Builder(LocalContext.current)
+                .data(imageUrl)
+                .crossfade(Constants.CROSSFADE_DURATION)
+                .build(),
+            contentDescription = null,
+            contentScale = ContentScale.Crop,
+            modifier = modifier,
+            alignment = Alignment.Center,
+            colorFilter = ColorFilter.tint(
+                color = tintColor,
+                blendMode = BlendMode.SrcAtop
+            )
+        )
+    } else {
+        Image(
+            painter = painterResource(R.drawable.background),
+            contentDescription = null,
+            contentScale = ContentScale.Crop,
+            modifier = modifier,
+            alignment = Alignment.TopCenter,
+            colorFilter = ColorFilter.tint(
+                color = tintColor,
+                blendMode = BlendMode.SrcAtop
+            )
+        )
+    }
+}
+
+@Composable
+fun MediaCharacters(
+    characters: List<Character>,
+    modifier: Modifier = Modifier,
+    contentPadding: PaddingValues = PaddingValues()
+) {
+    Column(modifier) {
+        Text(
+            text = stringResource(R.string.characters),
+            color = MaterialTheme.colorScheme.onBackground,
+            style = MaterialTheme.typography.titleMedium,
+            modifier = Modifier.padding(contentPadding)
+        )
+
+        Spacer(Modifier.size(dimensionResource(R.dimen.medium_padding)))
+
+        MediaSmallRow(
+            mediaList = characters
+        ) { character ->
+            MediaSmall(
+                image = character.image,
+                label = character.name,
+                onClick = { Log.d("CharacterId", "${character.id}") },
+                modifier = Modifier.width(dimensionResource(R.dimen.character_card_width))
+            )
+        }
+    }
+}
+
+@Composable
+fun MediaDetails(
+    title: String,
+    description: String,
+    modifier: Modifier = Modifier
+) {
+    Column(modifier) {
+        Text(
+            text = title,
+            color = MaterialTheme.colorScheme.onBackground,
+            style = MaterialTheme.typography.titleLarge,
+            maxLines = 4,
+            overflow = TextOverflow.Ellipsis
+        )
+
+        Text(
+            text = description,
+            color = MaterialTheme.colorScheme.onBackground.copy(
+                alpha = ContentAlpha.medium
+            ),
+            style = MaterialTheme.typography.bodyMedium,
+            modifier = Modifier
+                .verticalScroll(rememberScrollState())
+                .scrim(
+                    color = MaterialTheme.colorScheme.background,
+                    height = dimensionResource(R.dimen.small_padding)
+                )
+                .padding(vertical = dimensionResource(R.dimen.small_padding))
+        )
+    }
+}
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun MediaGenres(
+    genres: List<String>,
+    modifier: Modifier = Modifier,
+    contentPadding: PaddingValues,
+    color: Color = MaterialTheme.colorScheme.primaryContainer
+) {
+    LazyRow(
+        horizontalArrangement = Arrangement.spacedBy(
+            dimensionResource(R.dimen.medium_padding)
+        ),
+        contentPadding = contentPadding,
+        modifier = modifier
+    ) {
+        items(genres) { genre ->
+            SuggestionChip(
+                label = {
+                    Text(
+                        text = genre.lowercase(),
+                        color = MaterialTheme.colorScheme.onBackground,
+                        style = MaterialTheme.typography.labelMedium,
+                        modifier = Modifier.padding(
+                            vertical = dimensionResource(R.dimen.small_padding)
+                        )
+                    )
+                },
+                onClick = { },
+                shape = CircleShape,
+                colors = SuggestionChipDefaults.suggestionChipColors(
+                    containerColor = color.copy(alpha = 0.25f)
+                ),
+                border = SuggestionChipDefaults.suggestionChipBorder(
+                    borderColor = Color.Transparent
+                )
+            )
+        }
+    }
+}
+
+@Composable
+fun MediaStats(
+    stats: List<Stat>,
+    modifier: Modifier = Modifier
+) {
+    Row(
+        verticalAlignment = Alignment.CenterVertically,
+        horizontalArrangement = Arrangement.SpaceAround,
+        modifier = modifier
+    ) {
+        stats.forEach { stat ->
+            if (stat.label != StatLabel.UNKNOWN) {
+                Column(
+                    verticalArrangement = Arrangement.SpaceEvenly,
+                    horizontalAlignment = Alignment.CenterHorizontally
+                ) {
+                    Text(
+                        text = stat.label.value,
+                        color = MaterialTheme.colorScheme.onBackground,
+                        style = MaterialTheme.typography.labelSmall
+                    )
+
+                    Text(
+                        text = when (stat.label) {
+                            StatLabel.SCORE -> "${stat.score}%"
+                            StatLabel.RATING, StatLabel.POPULARITY -> "#${stat.score}"
+                            else -> ""
+                        },
+                        color = MaterialTheme.colorScheme.onBackground,
+                        style = MaterialTheme.typography.displaySmall
+                    )
+                }
+            }
+        }
+    }
+}
+
+@Composable
+fun MediaTrailer(
+    trailer: Trailer,
+    modifier: Modifier = Modifier
+) {
+    Column(modifier) {
+        Text(
+            text = stringResource(R.string.trailer),
+            color = MaterialTheme.colorScheme.onBackground,
+            style = MaterialTheme.typography.titleMedium
+        )
+
+        Spacer(Modifier.size(dimensionResource(R.dimen.medium_padding)))
+
+        val context = LocalContext.current
+        Box(
+            modifier = Modifier
+                .wrapContentSize()
+                .clip(RoundedCornerShape(dimensionResource(R.dimen.trailer_corner_radius)))
+                .clickable {
+                    val appIntent = Intent(Intent.ACTION_VIEW, Uri.parse(trailer.link))
+                    context.startActivity(appIntent)
+                }
+        ) {
+            AsyncImage(
+                model = ImageRequest.Builder(LocalContext.current)
+                    .data(trailer.thumbnail)
+                    .crossfade(Constants.CROSSFADE_DURATION)
+                    .build(),
+                contentDescription = stringResource(R.string.trailer),
+                contentScale = ContentScale.FillWidth,
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .aspectRatio(1.778f) // 16 : 9
+                    .clip(
+                        RoundedCornerShape(dimensionResource(R.dimen.trailer_corner_radius))
+                    ),
+                alignment = Alignment.Center
+            )
+
+            Image(
+                painter = painterResource(R.drawable.youtube),
+                contentDescription = null,
+                modifier = Modifier.align(Alignment.Center)
+            )
+        }
     }
 }
